Fixed issue with image cropping
Earlier, the data was 300x300 pixels, but were cropped to 224x224 pixels so they could be properly trained by the model. This cropping led to errors in the model. I retook pictures in 224x224 pixel size to ensure proper model training. keras_modelv3 classifies A, B, and C very well.